requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
Django==4.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
ipdb==0.13.11
ipython==8.8.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pure-eval==0.2.2
Pygments==2.14.0
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.8.1
tzdata==2022.7
wcwidth==0.2.5
